feat: centralize image embedding
